Update README to clarify Rails 5 section
I was updating the tutorial for the upcoming Rails Girls LA workshop.
We used to use Rails 4 now we are using Rails 5. I was wondering if we
still needed to use rails_12factor. The section on this in your README
was unclear to me. I spaced it out to clarify this section to how I
think it is supposed to read. I am not sure if it is right though.